Move five games' bot strategy out of gameplay.ts
AGENTS.md draws the line twice: "a predicate that exists only to characterise
the winning strategy belongs with the bot, not in gameplay.ts", and a spec is
"named after the module whose behaviour it asserts". Five games sat on the
wrong side of both — the whole search and both move-choosers lived in
gameplay.ts behind a two-line bot-strategy.ts wrapper, and their bot tests sat
in gameplay.spec.ts with no bot-strategy.spec.ts at all:

  four-piles-two-grabs   getSmartBotMove, getRandomBotMove
  latin-square-filling   optimalWinner, getSmartBotStep, getRandomBotStep
  two-of-three-takeaway  isWinningBoard, getSmartBotMove, getRandomBotMove
  three-piles-rebuild    splitWinningNumber, getLosingBotStep, get*BotStep
  sum-fifteen            winnerOptimal, chooseSmartMove, chooseTestMove

All five are ex-helpers.ts files, so this is the move that rename should have
made. Every other game in the repo is already shaped this way.

What stays in gameplay.ts is the win/loss predicate a generateStartBoard reads
to balance its start boards — the exemption AGENTS.md grants explicitly, and
the reason four-piles-two-grabs keeps isWinningBoard and three-piles-rebuild
keeps isWinningNumber/isLosingNumber. two-of-three-takeaway and sum-fifteen
build their start boards without consulting the characterisation, so theirs
move wholesale.

Tests move to match, and a handful of assertions move with the module whose
behaviour they were really about: latin-square-filling's "a completed square is
a first-player win", two-of-three-takeaway's immediate-win detection and its
generateStartBoard 50/50 balance check. The move also exposed that
sum-fifteen's test bot had no coverage at all — its opponent is a seeded RNG,
not chooseTestMove — so that gets the two cases the other games' test bots have.

Also drops the last ten references to a helpers module that no longer exists,
including two in game files pointing at a helpers.ts (sum-fifteen,
chocolate-breaking) that were never even in this group.

No behaviour changes: same functions, same call sites, 1588 tests green.
