+++ src/components/games/four-piles-two-grabs/four-piles-two-grabs.tsx
@@ -154,7 +154,7 @@ export const FourPilesTwoGrabs = strategyGameFactory({
       generateStartBoard,
       label: { hu: 'Teszt', en: 'Test' }
     },
-    // smart bot: verified as optimal (see helpers.spec.ts exhaustive minimax check)
+    // smart bot: verified as optimal (see bot-strategy.spec.ts exhaustive minimax check)
     {
       botStrategy: smartBotStrategy,
       generateStartBoard,
