+++ src/components/games/sum-fifteen/sum-fifteen.tsx
@@ -102,7 +102,7 @@ export const SumFifteen = strategyGameFactory({
       label: { hu: 'Teszt', en: 'Test' }
     },
     {
-      // smart bot: full optimal search (see winnerOptimal in helpers.ts)
+      // smart bot: full optimal search (see winnerOptimal in bot-strategy.ts)
       botStrategy: smartBotStrategy,
       generateStartBoard,
       label: { hu: 'Teljes', en: 'Full' },
