+++ src/components/games/two-of-three-takeaway/two-of-three-takeaway.tsx
@@ -121,7 +121,7 @@ export const TwoOfThreeTakeaway = strategyGameFactory({
       generateStartBoard,
       label: { hu: 'Teszt', en: 'Test' }
     },
-    // smart bot: verified as optimal (see helpers.spec.ts exhaustive minimax check)
+    // smart bot: verified as optimal (see bot-strategy.spec.ts exhaustive minimax check)
     {
       botStrategy: smartBotStrategy,
       generateStartBoard,
